Added Robert Cailliau www icon
We can revert this easily in the SVG if this doesn't work but I thought
it could be fun to try to put the web in the WACZ box, using Cailliau's
original (public domain) icon for the World Wide Web.

https://commons.wikimedia.org/wiki/File:WWW_logo_by_Robert_Cailliau.svg

Conceptually I like it because it includes not just HTML, but HTTP, URL,
etc.