Added todos and restructured outline
This commit includes many TODO sections for things that need to be
clarified or simplified. It also reorganizes the outline structure to
include Conformance, Terminology sections, and also adds a CDXJ section
for details about CDXJ (if we make it part of the WACZ specification). I
think other people could theoretically use CDXJ by itself if they
wanted by referencing the WACZ spec?