WACZ 1.1 ReSpec
This commit adds a v1.1.1 of the WACZ specification. The essential
features are derived from the [v1.1.1 README] and also from an
examination of the WACZ files that are created by browsertrix
v0.5.0-beta.8 which has been used heavily as part of the
[SUCHO](https://www.sucho.org) project.

In addition the the v1.2.0 version of the WACZ specification has been
flagged as a draft with `specStatus: Draft` in the configuration. A new
Changes section has been introduced to track changes from the previous
version. In order to do this a [Webrecorder profile for
ReSpec](https://github.com/webrecorder/respec) has been created. This is
a practice that the ReSpec project encourage for non-W3C entities to
reuse the spec framework.

The respec-webrecorder.js and WACZ logo have been placed in a shared
assets directory.

Closes #119