Reorganized specifications
This commit reorganizes the directory structure to allow multiple
specifications live in this repository. It assumes that the hostname
specs.webrecorder.net has been set up to point to the GitHub pages site
for the webrecorder/specs repository.

In addition to the WACZ and Use Cases specifications the WACZ Signing
and Verification specification has been moved here and turned into a
ReSpec document.